+++ packages/spotlight/Accelerator/Accelerator.js
@@ -63,7 +63,7 @@ class Accelerator {
 		 * Controls the frequency with which the acceleration will "freeze". While frozen,
 		 * the current target item cannot change, and all events are directed to it.
 		 *
-		 * @type {Number[]}
+		 * @type {Number|Number[]}
 		 * @default [3, 3, 3, 2, 2, 2, 1]
 		 * @public
 		 */
@@ -99,10 +99,14 @@ class Accelerator {
 						seconds = Math.floor(elapsedTime / 1000),
 						toSkip = 0;
 
-					seconds = seconds > this.frequency.length - 1 ? this.frequency.length - 1 : seconds;
+					if (Array.isArray(this.frequency)) {
+						seconds = seconds > this.frequency.length - 1 ? this.frequency.length - 1 : seconds;
+						toSkip = this.frequency[seconds] - 1;
+					} else {
+						toSkip = this.frequency - 1;
+					}
 
-					toSkip = this.frequency[seconds] - 1;
-					if (toSkip < 0) {
+					if (isNaN(toSkip) || toSkip < 0) {
 						toSkip = 0;
 					}